Address review: move suggestion-cache tests to a dedicated Swift Testing file
- Move BrowserHistorySuggestionCacheTests out of BrowserPanelTests.swift into
  its own pure Swift Testing file so the file no longer mixes Swift Testing with
  XCTestCase (cubic P3, per team policy). Wire the new file into the cmux-unit
  test target in project.pbxproj (PBXFileReference + PBXBuildFile + group +
  Sources phase); normalize-pbxproj and the test-wiring lint pass.
- Document the entries -> cachedSuggestionCandidates invalidation invariant at
  the didSet enforcement point, making the implicit contract explicit and
  guarding against a future bypassing writer (greptile P2).
